feat: Add ability to receive server broadcast messages - Messages will be visible as notification + notification dialog
fix: add some missing translation vars
- "Open_window" and "Do_you_really_want_to_dismiss_all_notices"
