+++ locales/en.json
@@ -388,7 +388,10 @@
     "muc_not-acceptable": "You have to use the reserved nickname.",
     "muc_registration-required": "You are not allowed to join this room, because you are not on the member list.",
     "muc_conflict": "Your nickname was already taken.",
-    "muc_service-unavailable": "You are not able to join, because the maximum number of users has been reached."
+    "muc_service-unavailable": "You are not able to join, because the maximum number of users has been reached.",
+    "broadcast" : "Broadcast",
+    "Open_window": "Open window",
+    "Do_you_really_want_to_dismiss_all_notices": "Do you really want to dismiss all notices?"
   },
   "Notifications": "Notifications"
 }
